:bug: Do not set keycloak hostname to docker container name!
Now we can run keycloak locally with Smile CDR and the admin console
works

Initially I thought we needed this bc the URLs at
when I visit the endpoint /.well-known/openid-configuration in the
browser they always point to localhost:8080. localhost:8080 is not recognized
inside the Smile CDR container, so I thought I needed to change
KC_HOSTNAME so all URLs point to `<keycloak container name>`:8080

Turns out we DO NOT NEED TO DO THIS! It works bc we configure the OIDC
server in Smile CDR with `http://keycloak:8080/realms/<realm name>`. The
way keycloak works is when the hostname settings are not set, the base URL for
these endpoints is based on the incoming request so that the
HTTP scheme, host, port, and path, are the same from the request.

See https://www.keycloak.org/server/hostname